add cardSizeScope to CardsExplorer
the cards explorer retrives the value based on the cardSizeScope
and provides the value and the update function to the children.
for now we dont store different scoped values in the context
but a context must be provided otherwise the view will fallback
to the default 270. cardsexplorer component also updates
the css variable accordingly.

- related to #63